+++ src/frequenz/sdk/actor/power_distributing/_distribution_algorithm/__init__.py
@@ -4,6 +4,7 @@
 """Utilities to manage power in a microgrid."""
 
 from ._distribution_algorithm import (
+    AggregatedBatteryData,
     DistributionAlgorithm,
     DistributionResult,
     InvBatPair,
@@ -13,4 +14,5 @@ __all__ = [
     "DistributionAlgorithm",
     "DistributionResult",
     "InvBatPair",
+    "AggregatedBatteryData",
 ]